+++ app/login/src/main/java/com/android/myfooddiarybookaos/login/mainSubUi/MidLayer.kt
@@ -20,6 +20,8 @@ import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.compose.ui.window.Dialog
+import androidx.compose.ui.window.Popup
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.android.myfooddiarybookaos.core.data.R
 import com.android.myfooddiarybookaos.data.robotoRegular
@@ -43,13 +45,16 @@ fun MidLayout(
         viewModel.goMain(LocalContext.current)
         viewModel.saveEmailState(LocalContext.current,emailText.value.text)
     }
+    var checkEnter by remember { mutableStateOf(0.3f) }
 
-    var checkEnter by remember {
-        mutableStateOf(0.3f)
-    }
+    val findPassPopState = remember { mutableStateOf(false) }
+    val isValid = remember { mutableStateOf(true) }
+    val loginFailCount = remember { mutableStateOf(0) }
 
-    val isValid = remember {
-        mutableStateOf(true)
+    if (loginFailCount.value  > 0 ){
+        isValid.value = false
+    } else if (loginFailCount.value == 5){
+        findPassPopState.value = true
     }
     Column(horizontalAlignment = Alignment.CenterHorizontally) {
         // EditText - email
@@ -67,12 +72,15 @@ fun MidLayout(
 
         Spacer(modifier = Modifier.height(4.dp))
         Text(
-            text = if (!isValid.value) "*아이디 또는 비밀번호를 잘못 입력했습니다. (n/5)"
+            text = if (!isValid.value) "*아이디 또는 비밀번호를 잘못 입력했습니다. (${loginFailCount.value}/5)"
             else "",
             color = colorResource(id = R.color.not_valid_text_color),
             fontFamily = robotoRegular,
             fontWeight = FontWeight(500),
-            fontSize = 12.scaledSp()
+            fontSize = 12.scaledSp(),
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(start = 16.dp)
         )
         Spacer(modifier = Modifier.height(7.dp))
 
@@ -88,7 +96,13 @@ fun MidLayout(
                         emailText.value.text,
                         pwText.value.text,
                         userState = {
-                            goMainResult = it
+                            if (it) {
+                                goMainResult = it
+                            } else {
+                                if (loginFailCount.value < 5) {
+                                    loginFailCount.value += 1
+                                }
+                            }
                         }
                     )
                 }
@@ -129,5 +143,15 @@ fun MidLayout(
         }
 
 
+    }
+
+    if (findPassPopState.value){
+        Dialog(
+            onDismissRequest = {
+                findPassPopState.value = false
+            }
+        ){
+
+        }
     }
 }
